add invalid source exception

+++ push-configs.ps1
@@ -11,26 +11,29 @@ $RepoTermpath = Join-Path -PATH $RepoPath -ChildPath "\wt\settings.json"
 
 $RepoPSpath = Join-Path -PATH $RepoPath -ChildPath "PowerShell\Microsoft.PowerShell_profile.ps1"
 
-function Push-Config {
+
+function Push-Config
+{
     param (
         $inputPath,
         $outputPath
     )
 
-    Write-Output "`nWriting config from $inputPath to $outputPath..."
-
-    if ( Test-Path $outputPath ) {
-        Remove-Item -PATH $outputPath -Recurse
-        Write-Host "Deleted existing config in " -ForegroundColor Red -NoNewline
-        Write-Host "$outputPath."
+    if (-Not(Test-Path $inputPath))
+    {
+        throw "$inputPath is not a valid repo path to copy from."
     }
-    else{
-        Write-Host "No existing config found, pushing..."
+
+    if ( Test-Path $outputPath )
+    {
+        Remove-Item -PATH $outputPath -Recurse
+        Write-Host "Deleted existing config in $outputPath." 
+    } else
+    {
+        Write-Host "No existing config found in $outputPath, pushing..."
     }
     Copy-Item -PATH $inputPath -Destination $outputPath -Recurse
-    Write-Host "Pushed config to " -ForegroundColor Green -NoNewline
-    Write-Host $outputPath -ForegroundColor White -NoNewline
-    Write-Host " successfully." -ForegroundColor green
+    Write-Host "Config push successful.`n" -ForegroundColor Green -NoNewline
 }
 
 Push-Config $RepoVimpath $WinVimpath
